sync patch r10783, r10787 from main trunk. Add security check to make code run safe.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/branches/UDK2010@11557 6f19259b-4bc3-4df7-8a09-765794883524

+++ UefiCpuPkg/CpuDxe/CpuDxe.c
@@ -1062,7 +1062,7 @@ InitInterruptDescriptorTable (
   IdtPtrAlignmentBuffer = AllocatePool (sizeof (*IdtPtr) + 16);
   IdtPtr = ALIGN_POINTER (IdtPtrAlignmentBuffer, 16);
   IdtPtr->Base = (UINT32)(((UINTN)(VOID*) gIdtTable) & (BASE_4GB-1));
-  IdtPtr->Limit = sizeof (gIdtTable) - 1;
+  IdtPtr->Limit = (UINT16) (sizeof (gIdtTable) - 1);
 
   AsmWriteIdtr (IdtPtr);
